+++ mvc/src/test/java/com/interface21/webmvc/servlet/mvc/HandlerExecutionHandlerAdapterTest.java
@@ -17,8 +17,7 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
 import com.interface21.webmvc.servlet.ModelAndView;
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerExecution;
+import com.interface21.webmvc.servlet.mvc.legacy.Controller;
 
 import samples.TestAnnotationController;
 import samples.TestExtendsController;
