+++ mvc/src/test/java/com/interface21/webmvc/servlet/mvc/HandlerMappingRegistryTest.java
@@ -12,8 +12,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerExecution;
-
 class HandlerMappingRegistryTest {
 
     @Nested
